dom/indexedDB/test/test_add_twice_failure.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<html>
<head>
  <title>Indexed Database Property Test</title>

  <script type="text/javascript" src="/MochiKit/packed.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

  <script type="text/javascript;version=1.7">
    function testSteps()
    {
      const CONSTRAINT_ERR =
        Components.interfaces.nsIIDBDatabaseException.CONSTRAINT_ERR;

      const name = window.location.pathname;
      const description = "My Test Database";

      let request = moz_indexedDB.open(name, description);
      request.onerror = errorHandler;
      request.onsuccess = grabEventAndContinueHandler;
      let event = yield;

      let db = event.result;

      request = db.setVersion("1");
      request.onerror = errorHandler;
      request.onsuccess = grabEventAndContinueHandler;
      let event = yield;

      let objectStore = db.createObjectStore("foo", "");
      let key = 10;

      request = objectStore.add({}, key);
      request.onerror = errorHandler;
      request.onsuccess = grabEventAndContinueHandler;
      event = yield;

      is(event.result, key, "Correct key");

      request = objectStore.add({}, key);
      request.onerror = new ExpectError(CONSTRAINT_ERR);
      request.onsuccess = unexpectedSuccessHandler;
      yield;

      finishTest();
      yield;
    }
  </script>
  <script type="text/javascript;version=1.7" src="helpers.js"></script>

</head>

<body onload="runTest();"></body>

</html>
